moved dashboard from home to /dashboard and creating a landing page (CTA at the very top to dashboard)
